setting up for final submit
I have moved all our final requirments to a holder. Subfolder files/enviroment from documents and recordings.

to the translator I have also added the lower case options for choice just in case a user inputs a lower or upper case letter we are looking for.  Friendly for both windows and linux.

A makefile was added to the enviroment for easily compilation and deconstruction, in linux/empress.